+++ src/connect/types/ui.ts
@@ -1,11 +1,15 @@
 import {
+    DocumentType,
+    DRIVE,
     driveLocations,
+    FILE,
+    FOLDER,
     nodeDropdownMenuOptions,
     nodeTypes,
     sharingTypes,
     syncStatuses,
 } from '@/connect';
-import { SynchronizationUnit } from 'document-model/document';
+import { Maybe, Scalars, SynchronizationUnit } from 'document-model/document';
 
 export type SharingTypes = typeof sharingTypes;
 export type NodeTypes = typeof nodeTypes;
@@ -14,32 +18,41 @@ export type SharingType = SharingTypes[number];
 export type DriveLocations = typeof driveLocations;
 export type DriveLocation = DriveLocations[number];
 
-type FileNode = {
+export type UiFileNode = {
+    kind: typeof FILE;
     id: string;
     name: string;
-    kind: 'file';
-    documentType: string;
+    documentType: DocumentType;
     parentFolder: string;
+    driveId: string;
+    syncStatus: SyncStatus | undefined;
     synchronizationUnits: SynchronizationUnit[];
 };
 
-type FolderNode = {
+export type UiFolderNode = {
+    kind: typeof FOLDER;
     id: string;
     name: string;
-    kind: 'folder';
     parentFolder: string;
-    children: (FileNode | FolderNode)[];
+    driveId: string;
+    children: UiNode[];
+    syncStatus: SyncStatus | undefined;
 };
 
-type DriveNode = {
+export type UiNode = UiDriveNode | UiFileNode | UiFolderNode;
+
+export type UiDriveNode = {
+    kind: typeof DRIVE;
     id: string;
     name: string;
-    kind: 'drive';
-    type: 'local' | 'cloud' | 'public';
-    sharingType: 'private' | 'public' | 'shared';
-    availableOffline: boolean;
     parentFolder: null;
-    children: (FileNode | FolderNode)[];
+    driveId: string;
+    children: UiNode[];
+    nodeMap: Record<string, UiNode>;
+    syncStatus: SyncStatus | undefined;
+    sharingType: SharingType;
+    availableOffline: boolean;
+    icon: string | null;
 };
 
 export type SyncStatuses = typeof syncStatuses;
@@ -49,3 +62,34 @@ export type SyncStatus = SyncStatuses[number];
 export type NodeDropdownMenuOptions = typeof nodeDropdownMenuOptions;
 
 export type NodeDropdownMenuOption = NodeDropdownMenuOptions[number];
+
+export type FileNode = {
+    documentType: Scalars['String']['output'];
+    id: Scalars['String']['output'];
+    kind: Scalars['String']['output'];
+    name: Scalars['String']['output'];
+    parentFolder: Maybe<Scalars['String']['output']>;
+    synchronizationUnits: SynchronizationUnit[];
+};
+
+export type FolderNode = {
+    id: Scalars['String']['output'];
+    kind: Scalars['String']['output'];
+    name: Scalars['String']['output'];
+    parentFolder: Maybe<Scalars['String']['output']>;
+};
+
+export type DocumentDriveDocument = {
+    state: {
+        global: {
+            id: string;
+            name: string;
+            icon: string;
+            nodes: (FileNode | FolderNode)[];
+        };
+        local: {
+            sharingType: SharingType;
+            availableOffline: boolean;
+        };
+    };
+};
